Create Create_Threat_Research_Analysis_VM
The files for the "Create_Threat_Research_Analysis_VM" Workshop are stored in this folder. The idea for this workshop is to help people semi-automatically build Threat Research Analysis VM using Open Source tools and projects created by others. The Threat Research Analysis VM is useful for conducting research for the following:
*Malware Anaylsis
*Threat Hunting
*Detection Engineering
*Forensic Analysis
*Incident Response